Add support for tracing connect()
Snitchrs needs a fast way to keep track of which pid is using which
connection.

One way would be to parse /proc tree from user space and map pid to
port. But this might loose track of connection and it's inefficient.

An alternative would be to keep track of all connect() and accept().
This commit adds support for commit. I've also committed some bindgen
bindings to get the sockaddr_in datastructure in rust - I shall include
that in the build process, but for now it's committed in the repo.

OFC, why bother with xdp/classifiers? I could just have tracked reads
and writes on the sockets fd. First point is that I wanted to try
classifiers. Also I'm not sure how easy and accurate it is to track it
attaching to syscalls. Therefore I'm not planning to rewrite the ebpf
for the write/read on the sockets.